style: figma plugin 탭 UI 추가
