+++ apps/figma-plugin/src/components/SemanticColorsTab.tsx
@@ -1,0 +1,13 @@
+export const SemanticColorsTab = () => {
+    return (
+        <>
+            <div className="flex items-center justify-center py-12">
+                <div className="text-center">
+                    <div className="text-gray-400 text-lg mb-2">🚧</div>
+                    <div className="text-sm text-gray-500">Coming Soon</div>
+                    <div className="text-xs text-gray-400 mt-1">Semantic color customization</div>
+                </div>
+            </div>
+        </>
+    );
+};
